refactor(host): fold the brain and memory wiring into the host
