+++ apps/desktop/src/main/brain/wiring.test.ts
@@ -1,9 +1,10 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import type { BrainStateRepository } from "@sidecar/brain";
+import { BRAIN_WAKE_KIND, type BrainStateRepository } from "@sidecar/brain";
 import { CREDENTIAL_REFERENCE_KIND, memoryChildStore } from "@sidecar/runtime";
 import { MAIN_SESSION_KEY, threadSessionKey } from "@sidecar/runtime-contracts";
-import { type BrainWiringDependencies, wireBrain } from "./wiring";
+import { normalizeSession, SESSION_STATUS, type Session } from "@sidecar/session";
+import { type BrainWiringDependencies, wakeEventsFromHooks, wireBrain } from "./wiring";
 
 /**
  * The brain wiring reaches a conversation's store only when a brain may
@@ -85,4 +86,49 @@ test("with no model to run on, a rebuild opens no conversation and loads no stor
   assert.equal(brains.current(threadSessionKey("t-1")), undefined);
   await brains.closeConversation(threadSessionKey("t-1"));
   brains.retire();
+});
+
+const NOW = 1_800_000_000_000;
+
+test("every hook event wakes the brain, carrying the session when the roster holds it", () => {
+  const held = normalizeSession(
+    { id: "claude-code", displayName: "Claude Code" },
+    {
+      providerSessionId: "session-a",
+      title: "Fix the flaky test",
+      status: SESSION_STATUS.COMPLETE,
+      lastActivityAt: NOW - 1_000,
+    },
+  );
+  const registry = {
+    get: (identity: { providerSessionId: string }): Session | undefined =>
+      identity.providerSessionId === "session-a" ? held : undefined,
+  };
+
+  const wakes = wakeEventsFromHooks(
+    "claude-code",
+    [
+      { providerSessionId: "session-a", event: "stop", atMs: NOW - 500 },
+      { providerSessionId: "session-b", event: "prompt", atMs: Number.NaN },
+    ],
+    registry,
+    NOW,
+  );
+
+  assert.equal(wakes.length, 2);
+  assert.deepEqual(wakes[0], {
+    kind: BRAIN_WAKE_KIND.HOOK,
+    identity: { providerId: "claude-code", providerSessionId: "session-a" },
+    hookEvent: "stop",
+    session: held,
+    atMs: NOW - 500,
+  });
+  // A hook for a session the poll has not seen yet still wakes the brain,
+  // dated now when the spool carried no usable time.
+  assert.deepEqual(wakes[1], {
+    kind: BRAIN_WAKE_KIND.HOOK,
+    identity: { providerId: "claude-code", providerSessionId: "session-b" },
+    hookEvent: "prompt",
+    atMs: NOW,
+  });
 });
